requirements.txt
psycopg2-binary
python-dotenv
pydantic
pytest
